Move geo util to their own package
