+++ composeApp/src/commonTest/kotlin/org/noiseplanet/noisecapture/signal/geojson/FeatureCollectionTest.kt
@@ -2,14 +2,14 @@ package org.noiseplanet.noisecapture.signal.geojson
 
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonPrimitive
-import org.noiseplanet.noisecapture.util.Feature
-import org.noiseplanet.noisecapture.util.FeatureCollection
-import org.noiseplanet.noisecapture.util.GeoJson
-import org.noiseplanet.noisecapture.util.LineString
-import org.noiseplanet.noisecapture.util.Point
-import org.noiseplanet.noisecapture.util.Polygon
-import org.noiseplanet.noisecapture.util.lineOf
-import org.noiseplanet.noisecapture.util.positionOf
+import org.noiseplanet.noisecapture.util.geo.Feature
+import org.noiseplanet.noisecapture.util.geo.FeatureCollection
+import org.noiseplanet.noisecapture.util.geo.GeoJson
+import org.noiseplanet.noisecapture.util.geo.LineString
+import org.noiseplanet.noisecapture.util.geo.Point
+import org.noiseplanet.noisecapture.util.geo.Polygon
+import org.noiseplanet.noisecapture.util.geo.lineOf
+import org.noiseplanet.noisecapture.util.geo.positionOf
 import kotlin.test.Test
 import kotlin.test.assertEquals
